% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverability.R
\name{clusterperm.glmer}
\alias{clusterperm.glmer}
\title{Cluster-based permutation tests for time series data, based on generalized linear mixed-effects models or other \code{buildmer} models. This is an alias for \code{clusterperm.lmer} provided for discoverability.}
\usage{
clusterperm.glmer(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{clusterperm.lmer}.}
}
\description{
Cluster-based permutation tests for time series data, based on generalized linear mixed-effects models or other \code{buildmer} models. This is an alias for \code{clusterperm.lmer} provided for discoverability.
}
\examples{
# Testing a single EEG electrode, with random effects by participants
\donttest{
perms <- clusterperm.glmer(Fz ~ dev*session + (dev*session|ppn),data=MMN,series.var=~time)
}
\dontshow{
perms <- clusterperm.glmer(Fz ~ dev*session + (1|ppn),data=MMN[MMN$time > 200 & MMN$time < 205,],series.var=~time,nperm=list(nsim=2),type='anova')
perms <- clusterperm.glmer(Fz ~ dev*session + (1|ppn),data=MMN[MMN$time > 200 & MMN$time < 205,],series.var=~time,nperm=list(nsim=2),type='regression')
perms <- clusterperm.glmer(Fz ~ session*cond + (1|ppn),data=within(MMN[MMN$time > 200 & MMN$time < 205,],{session <- factor(session); cond <- factor(cond)}),series.var=~time,nperm=list(nsim=2),type='regression')
}
}
\seealso{
clusterperm.lmer
}
