% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMN}
\alias{MMN}
\title{ERP data from Jager (in prep.). See the vignettes for details.}
\format{
A standard data frame.
}
\usage{
data(MMN)
}
\description{
ERP data from Jager (in prep.). See the vignettes for details.
}
\keyword{datasets}
