% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverability.R
\name{perm.glm}
\alias{perm.glm}
\title{A general permutation test for mixed-effects models or other \code{buildmer} models. This is an alias for \code{perm.lm} provided for discoverability.}
\usage{
perm.glm(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{perm.lm}.}
}
\description{
A general permutation test for mixed-effects models or other \code{buildmer} models. This is an alias for \code{perm.lm} provided for discoverability.
}
\examples{
\donttest{perms <- perm.glm(Fz ~ Deviant * Session,data=MMN)}
\dontshow{perms <- perm.glm(Fz ~ Deviant * Session,data=MMN[MMN$Time > 200 & MMN$Time < 205,],nperm=2)}
}
\seealso{
perm.lm, perm.lmer
}
