% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutes.R
\name{permu.plot}
\alias{permu.plot}
\title{Create a heatmap of the results of permutation testing}
\usage{
permu.plot(data)
}
\arguments{
\item{data}{Output of permu.test. You may want to subset it if you want to simulate zooming in.}
}
\value{
A ggplot2 object containing a heatmap of p-values. You may want to add a `scale_x_continuous(expand=c(0,0),breaks=(...))' to it to increase the granularity of the spacing to your satisfaction.
}
\description{
Create a heatmap of the results of permutation testing
}
