% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverability.R
\name{clusterperm.glm}
\alias{clusterperm.glm}
\title{Cluster-based permutation tests for time series data, based on generalized linear models or other \code{buildmer} models. This is an alias for \code{clusterperm.lmer} provided for discoverability.}
\usage{
clusterperm.glm(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{clusterperm.lm}.}
}
\description{
Cluster-based permutation tests for time series data, based on generalized linear models or other \code{buildmer} models. This is an alias for \code{clusterperm.lmer} provided for discoverability.
}
\examples{
perms <- clusterperm.glm(Fz ~ dev*session,data=MMN,series.var=~time)
}
\seealso{
clusterperm.lm, clusterperm.lmer
}
